Big update
Version 12?
Fixed incorrect data showing when an invalid effect is configured.
Fixed maps not getting renamed.
Fixed wanderer trades not getting added if "removeOtherTrades" = false.
Reworded error message to clarify that problem files should be placed through a JSON verifier AS WELL AS checked manually.

Added capability to force an npc's levels up. To do this:
	- You must be in creative
	- Open the villager's trade menu
	- Press the UP arrow on your keyboard
	- Exit the trade menu.
This will force the villager's xp to the required level, and then trigger the top most trade (I can't find a way to do this without consuming a trade).
This will not work in survival.
There is no "level down" functionality.

Added ability to colour custom potions and arrows. How to use:
- In the item object, create 3 new fields called:
	- "customColourR"
	- "customColourG"
	- "customColourB"
- (Please note that this uses the British spelling, with 'U' in the word)
- Apply respective RGB values, minimum is 0 and maximum is 255.
- These custom colours will only work with tipped arrows and potions.

Added NBT data array field. How to use:
- In the item object, create a new array called "nbtTags".
- Inside of the array, for each NBT tag you want to add, create a new object with the following fields:
	- "nbtName": The name of your NBT tag.
	- "dataType": The type of data it should be converted to. Possible values: String, Boolean, Integer, Float, Byte, Long
	- "data": The data to go with the NBT tag. This MUST be a string (it is converted in the code using the "dataType" field above.
Please note that custom nbt data will NOT work for potions or tipped arrows - this is because modifying the NBT can break that potion effect.